Fixed the proposed issues on Reports
