+++ app/controllers/reports_controller.rb
@@ -2,7 +2,7 @@ class ReportsController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    @report = Report.create
+    @report = Report.new
     @test_data = @report.admin_chart_data
     render 'report'
   end
