+++ app/models/story_point.rb
@@ -8,7 +8,4 @@ class StoryPoint < ApplicationRecord
   validates :sprint, presence: true
   validates :user, presence: true
 
-  scope :by_user_on_sprint, ->(user, sprint) do
-    where(user: user, sprint: sprint).take
-  end
 end
