+++ config/initializers/lib_require.rb
@@ -1,1 +1,2 @@
 require 'date_parser'
+require 'report'
