+++ db/migrate/20170908142018_drop_reports.rb
@@ -1,0 +1,5 @@
+class DropReports < ActiveRecord::Migration[5.0]
+  def change
+    drop_table :reports
+  end
+end
